refactor(l5): move read-policy from src/ into server/domain (issue 864)
src/read-policy.ts -> server/domain/read-policy.ts. The old path stays as a
re-export shim so existing importers keep working unchanged; the shim holds
no implementation, which is what the M2 rule asks for.

Shim closure moves 42 -> 41: src/read-policy.ts is now judged a shim rather
than an implementation file by scripts/done-means/864-moved-out-of-src.sh.

src/read-policy.test.ts is left in place under src/ per the M1 exception for
tests. It imports through the shim and passes unchanged.

Extraction is deferred to a later adapter lane once the adapter clause lands:
backgroundExtract takes 6 parameters against a max-params of 4, and five
legacy callers pass them positionally, so moving it needs the adapter shape
rather than a straight relocation.
